requirements.txt
fastapi
uvicorn
pandas
numpy~=1.19.5
sqlalchemy==1.4.47
pymysql
requests>=2.22.0
urllib3>=1.24.2
mysqlclient
PyYAML
python-multipart
pydantic
loguru

pandasql
jinja2